requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.17

# Database
motor>=3.6.0
pymongo>=4.10.0

# AI/LLM
openai>=1.55.0

# HTTP Client
httpx>=0.28.0

# RSS Parsing
feedparser>=6.0.11
beautifulsoup4>=4.12.3

# Task Scheduling
apscheduler>=3.10.4

# Configuration
python-dotenv>=1.0.1
pydantic-settings>=2.6.0

# Utilities
python-dateutil>=2.9.0

# Testing (dev dependencies)
pytest>=8.0.0
pytest-asyncio>=0.24.0
